requirements.txt
Django >= 1.9
django-mptt >= 0.8
